+++ src/lib.rs
@@ -102,12 +102,17 @@
 //! ## `async` support
 //! Since Rust at the time of writing does not natively support async methods in traits, you may opt in to having `#[async_trait]` generated for your trait:
 //!
-//! ```rust
-//! # use entrait::entrait;
-//! #[entrait(Foo, async_trait)]
-//! async fn foo<D>(deps: &D) {
-//! }
-//! ```
+#![cfg_attr(
+    feature = "async-trait",
+    doc = r##"
+```rust
+# use entrait::entrait;
+#[entrait(Foo, async_trait)]
+async fn foo<D>(deps: &D) {
+}
+```
+"##
+)]
 //! This is designed to be forwards compatible with real async fn in traits.
 //! When that day comes, you should be able to just remove the `async_trait=true` to get a proper zero-cost future.
 //!
@@ -135,12 +140,17 @@
 //! can to co-exist with macros like these. Since `entrait` is a higher-level macro that does not touch fn bodies (it does not even try to parse them),
 //! entrait should be processed after, which means it should be placed _before_ lower level macros. Example:
 //!
-//! ```rust
-//! # use entrait::entrait;
-//! #[entrait(FetchThing, no_deps, async_trait)]
-//! #[feignhttp::get("https://my.api.org/api/{param}")]
-//! async fn fetch_thing(#[path] param: String) -> feignhttp::Result<String> {}
-//! ```
+#![cfg_attr(
+    feature = "use-async-trait",
+    doc = r##"
+```rust
+# use entrait::entrait;
+#[entrait(FetchThing, no_deps)]
+#[feignhttp::get("https://my.api.org/api/{param}")]
+async fn fetch_thing(#[path] param: String) -> feignhttp::Result<String> {}
+```
+"##
+)]
 //!
 //! Here we had to use the `no_deps` entrait option.
 //! This is used to tell entrait that the function does not have a `deps` parameter as its first input.
@@ -419,14 +429,38 @@ assert_eq!("Hello World!", hello_string);
 
 #[cfg(feature = "unimock")]
 mod macros {
+    #[cfg(feature = "use-async-trait")]
+    pub use entrait_macros::entrait_export_unimock_use_async_trait as entrait_export;
+    #[cfg(feature = "use-async-trait")]
+    pub use entrait_macros::entrait_unimock_use_async_trait as entrait;
+
+    #[cfg(all(feature = "use-associated-future", not(feature = "use-async-trait")))]
+    pub use entrait_macros::entrait_export_unimock_use_associated_future as entrait_export;
+    #[cfg(all(feature = "use-associated-future", not(feature = "use-async-trait")))]
+    pub use entrait_macros::entrait_unimock_use_associated_future as entrait;
+
+    #[cfg(not(any(feature = "use-async-trait", feature = "use-associated-future")))]
     pub use entrait_macros::entrait_export_unimock as entrait_export;
+    #[cfg(not(any(feature = "use-async-trait", feature = "use-associated-future")))]
     pub use entrait_macros::entrait_unimock as entrait;
 }
 
 #[cfg(not(feature = "unimock"))]
 mod macros {
-    pub use entrait_macros::entrait;
+    #[cfg(feature = "use-async-trait")]
+    pub use entrait_macros::entrait_export_use_async_trait as entrait_export;
+    #[cfg(feature = "use-async-trait")]
+    pub use entrait_macros::entrait_use_async_trait as entrait;
+
+    #[cfg(all(feature = "use-associated-future", not(feature = "use-async-trait")))]
+    pub use entrait_macros::entrait_export_use_associated_future as entrait_export;
+    #[cfg(all(feature = "use-associated-future", not(feature = "use-async-trait")))]
+    pub use entrait_macros::entrait_use_associated_future as entrait;
+
+    #[cfg(not(any(feature = "use-async-trait", feature = "use-associated-future")))]
     pub use entrait_macros::entrait_export;
+    #[cfg(not(any(feature = "use-async-trait", feature = "use-associated-future")))]
+    pub use entrait_macros::entrait;
 }
 
 /// The entrait macro.
@@ -437,17 +471,12 @@ pub use macros::entrait_export;
 pub use ::implementation::Impl;
 
 /// Optional mock re-exports for macros
+#[cfg(feature = "unimock")]
 #[doc(hidden)]
-pub mod __m {
-    /// Re-export of the unimock crate.
-    #[cfg(feature = "unimock")]
-    pub mod unimock {
-        pub use ::unimock::*;
-    }
+pub use ::unimock as __unimock;
 
-    /// Re-export of the mockall crate.
-    #[cfg(feature = "mockall")]
-    pub mod mockall {
-        pub use ::mockall::*;
-    }
+#[cfg(feature = "async-trait")]
+#[doc(hidden)]
+pub mod __async_trait {
+    pub use ::async_trait::async_trait;
 }